sync-github-releases: give the target-commitish decision its own module
apply-release-plan.ts mixed two abstractions: walking the plan (create →
update → delete, dry-run gate, logging) and deciding which commit a
to-be-created release is tagged at. The latter is the trickiest domain logic
in the module — it carries the most spec coverage — yet a reader of "how do we
apply a plan" had to wade through it.

Move it to its own sync/target-commitish.ts (with its own spec), so each file
is one abstraction. Rename the confusingly-near pair while at it:
- resolveTargetCommitish: the impure entry that gathers the git facts (was
  resolveCreateTarget)
- decideTargetCommitish: the pure decision over those facts (was
  resolveTargetCommitish)

"resolve" now consistently means "do the git IO", "decide" means "pure
branching over given facts", so the two no longer read alike.
